ISS90 - refactor scenario visualizer
