+++ python_src/visualization/__init__.py
@@ -1,2 +1,3 @@
 from .experiment import ExperimentVisualizer
 from .mutation import Mutation
+from .scenario import Scenario
