+++ python_src/visualization/scenario.py
@@ -1,59 +1,11 @@
 import json
 from models import ac3r, ac3rp
 from matplotlib import pyplot as plt
-import matplotlib.patches as patches
-from shapely.geometry import LineString, Polygon
-from shapely.affinity import translate, rotate
+from shapely.geometry import LineString
 from descartes import PolygonPatch
-from math import atan2, pi, degrees
-import numpy as np
-import seaborn as sns
-import pandas as pd
-from models import SimulationFactory, Simulation, SimulationScore
-from models.ac3rp import CrashScenario
 
 
-#
-# # https://stackoverflow.com/questions/34764535/why-cant-matplotlib-plot-in-a-different-thread
-# class RoadTestVisualizer:
-#     """
-#         Visualize and Plot RoadTests
-#     """
-#
-#     little_triangle = Polygon([(10, 0), (0, -5), (0, 5), (10, 0)])
-#     square = Polygon([(5, 5), (5, -5), (-5, -5), (-5, 5), (5, 5)])
-#
-#
-
-
-class RoadVisualizer:
-
-    def __init__(self, the_figure):
-        self.the_figure = the_figure
-
-    def visualize(self, road):
-        # Be sure we plot on the right figure
-        plt.figure(self.the_figure.number)
-
-        # Road Geometry.
-
-        # TODO Assume the road has constant width.
-        road_width = road.road_nodes[0][3]
-        road_poly = LineString([(t[0], t[1]) for t in road.road_nodes]).buffer(road_width, cap_style=2, join_style=2)
-        road_patch = PolygonPatch(road_poly, fc='gray', ec='dimgray')  # ec='#555555', alpha=0.5, zorder=4)
-        plt.gca().add_patch(road_patch)
-
-        # Central line
-        # TODO Lane/configurations
-        sx = [t[0] for t in road.road_nodes]
-        sy = [t[1] for t in road.road_nodes]
-        plt.plot(sx, sy, 'yellow')
-
-        plt.draw()
-        plt.pause(0.001)
-
-
-class VehicleTrajectoryVisualizer:
+class Scenario:
 
     def __init__(self, the_figure):
         self.the_figure = the_figure
@@ -119,32 +71,3 @@ class VehicleTrajectoryVisualizer:
         plt.title(f'AC3R Plus {ac3rp_scenario.name}')
         plt.show()
         fig.savefig(f'data/{ac3rp_scenario.name}_ac3rp.png', bbox_inches="tight")
-
-
-class CrashScenarioVisualizer:
-
-    def __init__(self, map_size):
-        self.map_size = map_size
-        self.the_figure = None
-
-        # Make sure there's a windows and does not block anything when calling show
-        plt.ion()
-        plt.show()
-
-        # Setup the figure environment
-        self._setup_figure()
-
-    def _setup_figure(self):
-        if self.the_figure is None:
-            self.the_figure = plt.figure()
-        else:
-            plt.figure(self.the_figure.number)
-            plt.clf()
-        plt.gcf().set_title("Scenario Visualizer")
-        plt.gca().set_aspect('equal', 'box')
-        plt.gca().set(xlim=(-30, self.map_size + 30), ylim=(-30, self.map_size + 30))
-
-    def visualize_scenario(self, crash_scenario):
-        self._setup_figure()
-
-        # Visualize the road
